Finally have blogs sorted by year and category on the page-blog.php template.
However I think now that I know what I want the behavior to be, I may eventually scrap all of this some day and rewrite it so that it's scalable. I don't think I can easily add pagination using AJAX to the page with the way this is set up now.