Instead of echoing the script contents into the <head>, we are now referencing the script file before the closing </body>
I wanted to echo the contents of the script before the closing </body> but that doesn't seem possible. It seems that the best way forward is to simply enqueue the script before the closing </body>. I don't like this because it adds http requests but perhaps I am overthinking and this doesn't matter that much. We will keep the inline CSS in the <head> tag though for now. Until it proves to be problematic. I hope it doesn't though because I like the idea of keeping my core CSS separate from my ACF Block CSS in an effort to make the theme more modular.