Added 'Image Slideshow' ACF Block's script and styles to 404.php
I wish I could just edit this page like I do any other page so that I could use the_content() and then place my 'Image Slideshow' ACF Block within the Gutenberg editor... but that doesn't seem possible. So until I figure out how to do that we're going to just hardcode the slideshow HTML in to 404.php and then we're going to add code to functions.php that checks if the page is the 404 page and then adds the necessary styles and scripts.