Renamed prism.js to code.js and moved it in to the 'blocks/code' directory where the gulpfile.js can now watch it, and where functions.php can now find its js file for enqueueing and its css file for echoing inline - only when a 'code' element is detected
However there is a problem. I was using the HTML block previously because that let me add a class to the code block to get the desired synax highlighting from prism. This might be problematic because what if I want to use the HTML block for something else in the future? I think I need to go  back to every post and manually change them from HTML blocks to Code blocks but still keep the class intact. Perhaps I can just use the "Additional Custom Class(es)" field to add the class?